mcp-cursor/requirements.txt
mcp>=1.21.0
oracledb>=2.0.0
python-dotenv>=1.0.0

